style(flex7): Recreate the Holy Grail layout
While I did not use flex for centering the text in the header and footer
containers, I did learn how to vertically center texts using the height
and line-height properties.

Overall, my solution was close enough this time, however, I may need to
review this series of lessons for a better grasp.